src/com/cedarsolutions/santa/client/admin/AdminHistoryConverter.java
/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
 *
 *              C E D A R
 *          S O L U T I O N S       "Software done right."
 *           S O F T W A R E
 *
 * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
 *
 * Copyright (c) 2011-2012 Kenneth J. Pronovici.
 * All rights reserved.
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the Apache License, Version 2.0.
 * See LICENSE for more information about the licensing terms.
 *
 * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
 *
 * Author   : Kenneth J. Pronovici <pronovic@ieee.org>
 * Language : Java 6
 * Project  : Secret Santa Exchange
 *
 * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
package com.cedarsolutions.santa.client.admin;

import com.mvp4g.client.annotation.History;
import com.mvp4g.client.annotation.History.HistoryConverterType;
import com.mvp4g.client.history.HistoryConverter;

/**
 * History converter for the admin module.
 * @author Kenneth J. Pronovici <pronovic@ieee.org>
 */
@History(type = HistoryConverterType.NONE)
public class AdminHistoryConverter implements HistoryConverter<AdminEventBus> {

    /** This portion of the site cannot be crawled. */
    @Override
    public boolean isCrawlable() {
        return false;
    }

    /** Convert an event type and token into an event on the event bus. */
    @Override
    public void convertFromToken(String eventType, String param, AdminEventBus eventBus) {
        eventBus.dispatch(eventType);
    }

}
